issue-521: updated categories
