+++ docs/api/tests.services.md
@@ -1,0 +1,13 @@
+<!--
+  ~ Copyright (c) 2023-2024 Arista Networks, Inc.
+  ~ Use of this source code is governed by the Apache License 2.0
+  ~ that can be found in the LICENSE file.
+  -->
+
+# ANTA catalog for services tests
+
+::: anta.tests.services
+    options:
+      show_root_heading: false
+      show_root_toc_entry: false
+      merge_init_into_class: false
