+++ examples/tests.yaml
@@ -205,6 +205,7 @@ anta.tests.security:
           common_name: Arista Networks Internal IT Root Cert Authority
           encryption_algorithm: RSA
           key_size: 4096
+anta.tests.services:
   - VerifyHostname:
       hostname: s1-spine1
