+++ tests/units/anta_tests/test_services.py
@@ -1,0 +1,32 @@
+# Copyright (c) 2023-2024 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
+"""
+Tests for anta.tests.security.py
+"""
+from __future__ import annotations
+
+from typing import Any
+
+from anta.tests.services import VerifyHostname
+from tests.lib.anta import test  # noqa: F401; pylint: disable=W0611
+
+DATA: list[dict[str, Any]] = [
+    {
+        "name": "success",
+        "test": VerifyHostname,
+        "eos_data": [{"hostname": "s1-spine1", "fqdn": "s1-spine1.fun.aristanetworks.com"}],
+        "inputs": {"hostname": "s1-spine1"},
+        "expected": {"result": "success"},
+    },
+    {
+        "name": "failure-incorrect-hostname",
+        "test": VerifyHostname,
+        "eos_data": [{"hostname": "s1-spine2", "fqdn": "s1-spine1.fun.aristanetworks.com"}],
+        "inputs": {"hostname": "s1-spine1"},
+        "expected": {
+            "result": "failure",
+            "messages": ["Expected `s1-spine1` as the hostname, but found `s1-spine2` instead."],
+        },
+    },
+]
